SAK-41371: Problem using facelets in JSF2 with ViewHandler
