+++ jsf2/jsf2-app/src/META-INF/faces-config.xml
@@ -22,7 +22,7 @@
 
 	<application>
 		<view-handler>
-			org.sakaiproject.jsf2.app.SakaiViewHandler
+			org.sakaiproject.jsf2.app.SakaiViewHandlerWrapper
 		</view-handler>
 		<navigation-handler>
 			org.sakaiproject.jsf2.app.SakaiNavigationHandler
